Added navigation drawer implemented fragments into the drawer completed home fragment
