+++ app/src/main/java/com/teamzmron/selfstudyapp/Fragments/Adjective/AdjectiveHomeFragment.kt
@@ -1,4 +1,4 @@
-package com.teamzmron.selfstudyapp.Fragments
+package com.teamzmron.selfstudyapp.Fragments.Adjective
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,13 +8,15 @@ import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.teamzmron.selfstudyapp.Adapters.NounAdapter
-import com.teamzmron.selfstudyapp.Helper.NounSwipeToDeleteHelper
+import com.teamzmron.selfstudyapp.Adapters.AdjectiveAdapter
+import com.teamzmron.selfstudyapp.Helper.AdjectiveSwipeToDeleteHelper
 
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.ViewModel.NounViewModel
+import com.teamzmron.selfstudyapp.ViewModel.AdjectiveViewModel
+import com.teamzmron.selfstudyapp.ViewModel.ViewModelProviderFactory
 import dagger.android.support.DaggerFragment
-import kotlinx.android.synthetic.main.fragment_noun_home.*
+import kotlinx.android.synthetic.main.fragment_adjective_home.*
+import javax.inject.Inject
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -23,16 +25,22 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [NounHomeFragment.newInstance] factory method to
+ * Use the [AdjectiveHomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class NounHomeFragment : DaggerFragment(), NounAdapter.OnNounClickListener {
+class AdjectiveHomeFragment : DaggerFragment(), AdjectiveAdapter.OnAdjectiveClickListener {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var nounViewModel: NounViewModel
-    private lateinit var nounAdapter : NounAdapter
+    private lateinit var adjectiveViewModel: AdjectiveViewModel
+
+    @Inject
+    lateinit var adapter: AdjectiveAdapter
+
+    @Inject
+    lateinit var providerFactory: ViewModelProviderFactory
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -47,36 +55,33 @@ class NounHomeFragment : DaggerFragment(), NounAdapter.OnNounClickListener {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_noun_home, container, false)
+        return inflater.inflate(R.layout.fragment_adjective_home, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         initViewModels()
         initRecyclerView()
-
     }
 
     private fun initViewModels() {
-        nounViewModel = ViewModelProvider(this).get(NounViewModel::class.java)
+        adjectiveViewModel = ViewModelProvider(this, providerFactory).get(AdjectiveViewModel::class.java)
     }
 
     private fun initRecyclerView() {
-        nounAdapter = NounAdapter(context!!, nounViewModel, viewLifecycleOwner, this )
-        recycler_nounhome.layoutManager = LinearLayoutManager(context!!)
-        recycler_nounhome.adapter = nounAdapter
+        recycler_adjhome.layoutManager = LinearLayoutManager(context)
+        recycler_adjhome.adapter = adapter
 
         var itemTouchHelper = ItemTouchHelper(
-            NounSwipeToDeleteHelper(
+            AdjectiveSwipeToDeleteHelper(
                 this,
-                nounViewModel,
-                nounAdapter, 0, ItemTouchHelper.RIGHT
+                adjectiveViewModel,
+                adapter, 0, ItemTouchHelper.RIGHT
             )
         )
 
-        itemTouchHelper.attachToRecyclerView(recycler_nounhome)
-        nounAdapter.notifyDataSetChanged()
-
+        itemTouchHelper.attachToRecyclerView(recycler_adjhome)
+        adapter.notifyDataSetChanged()
     }
 
     companion object {
@@ -86,12 +91,13 @@ class NounHomeFragment : DaggerFragment(), NounAdapter.OnNounClickListener {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment NounListFragment.
+         * @return A new instance of fragment AdjectiveListFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            NounHomeFragment().apply {
+            AdjectiveHomeFragment()
+                .apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
@@ -99,7 +105,7 @@ class NounHomeFragment : DaggerFragment(), NounAdapter.OnNounClickListener {
             }
     }
 
-    override fun onNounClick(id: Int) {
-        TODO("Not yet implemented")
+    override fun onAdjectiveClick(id: Int) {
+
     }
 }
