+++ app/src/main/java/com/teamzmron/selfstudyapp/Fragments/Home.kt
@@ -2,6 +2,7 @@ package com.teamzmron.selfstudyapp.Fragments
 
 
 import android.app.AlertDialog
+import android.content.Context
 import android.os.Bundle
 import android.view.*
 import androidx.appcompat.app.AppCompatActivity
@@ -10,27 +11,31 @@ import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.ViewModelProvider
 import com.teamzmron.selfstudyapp.Adapters.NounAdapter
 import com.teamzmron.selfstudyapp.Adapters.ViewPagerAdapter
+import com.teamzmron.selfstudyapp.Fragments.Adjective.AdjectiveHomeFragment
+import com.teamzmron.selfstudyapp.Fragments.noun.NounHomeFragment
+import com.teamzmron.selfstudyapp.Fragments.verb.VerbHomeFragment
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.ViewModel.NounViewModel
-import com.teamzmron.selfstudyapp.ViewModel.PageViewModel
-import com.teamzmron.selfstudyapp.ViewModel.WordDetailsViewModel
-import com.teamzmron.selfstudyapp.ViewModel.WordViewModel
+import com.teamzmron.selfstudyapp.ViewModel.*
+import dagger.android.support.DaggerFragment
 import kotlinx.android.synthetic.main.fragment_home.*
+import javax.inject.Inject
 
 
 /**
  * A simple [Fragment] subclass.
  */
-class Home : Fragment(), NounAdapter.OnNounClickListener {
+class Home : DaggerFragment(), NounAdapter.OnNounClickListener {
 
     private lateinit var nounViewModel: NounViewModel
-    private lateinit var pageViewModel: PageViewModel
     private lateinit var wordViewModel: WordViewModel
-    private val wordDetailsViewModel: WordDetailsViewModel by activityViewModels()
+
 
     private val NOUN = "Noun"
     private val VERB = "Verb"
     private val ADJ = "Adjective"
+
+    @Inject
+    lateinit var providerFactory: ViewModelProviderFactory
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -66,9 +71,8 @@ class Home : Fragment(), NounAdapter.OnNounClickListener {
 
 
     private fun initViewModels() {
-        nounViewModel = ViewModelProvider(this).get(NounViewModel::class.java)
-        pageViewModel = ViewModelProvider(this).get(PageViewModel::class.java)
-        wordViewModel = ViewModelProvider(this).get(WordViewModel::class.java)
+        nounViewModel = ViewModelProvider(this, providerFactory).get(NounViewModel::class.java)
+        wordViewModel = ViewModelProvider(this, providerFactory).get(WordViewModel::class.java)
     }
 
 
@@ -80,23 +84,15 @@ class Home : Fragment(), NounAdapter.OnNounClickListener {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
             R.id.home -> {
-                if(fragmentManager!!.findFragmentByTag("addWord") != null) {
-                    removeFragment(AddWord())
-                }
-                pageViewModel.getFragmentTransaction(context!!)
-                    .replace(pageViewModel.getContainer(), Home(), "Home")
-                    .commit()
+
                 true
             }
             R.id.add -> {
-                if(fragmentManager!!.findFragmentByTag("addWord") == null) {
-                    addFragment(AddWord(),"addWord")
-                }
+
 
                 true
             }
             R.id.quiz -> {
-                addFragment(QuizSettings(), "quizSettings")
                 true
             }
             R.id.clearAllNoun -> {
@@ -119,17 +115,6 @@ class Home : Fragment(), NounAdapter.OnNounClickListener {
         }
     }
 
-    private fun addFragment(fragment: Fragment, tag: String) {
-        pageViewModel.getFragmentTransaction(context!!)
-            .add(pageViewModel.getContainer(),fragment,tag)
-            .commit()
-    }
-
-    private fun removeFragment(fragment: Fragment) {
-        pageViewModel.getFragmentTransaction(context!!)
-            .remove( fragment)
-            .commit()
-    }
 
 
     private fun promptBeforeClearingWords(category: String) {
@@ -166,10 +151,10 @@ class Home : Fragment(), NounAdapter.OnNounClickListener {
 
 
     override fun onNounClick(id: Int) {
-        wordDetailsViewModel.setMutableId(id)
-        pageViewModel.getFragmentTransaction(context!!)
-            .add(R.id.fragment_container, WordDetails(), "WordDetails")
-            .commit()
+
+//        pageViewModel.getFragmentTransaction(context!!)
+//            .add(R.id.fragment_container, WordDetails(), "WordDetails")
+//            .commit()
     }
