+++ app/src/main/java/com/teamzmron/selfstudyapp/Fragments/noun/NounHomeFragment.kt
@@ -1,19 +1,22 @@
-package com.teamzmron.selfstudyapp.Fragments
+package com.teamzmron.selfstudyapp.Fragments.noun
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.teamzmron.selfstudyapp.Adapters.VerbAdapter
-import com.teamzmron.selfstudyapp.Helper.VerbSwipeToDeleteHelper
-import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.ViewModel.VerbViewModel
-import kotlinx.android.synthetic.main.fragment_verb_home.*
+import com.teamzmron.selfstudyapp.Adapters.NounAdapter
+import com.teamzmron.selfstudyapp.Helper.NounSwipeToDeleteHelper
 
+import com.teamzmron.selfstudyapp.R
+import com.teamzmron.selfstudyapp.ViewModel.NounViewModel
+import com.teamzmron.selfstudyapp.ViewModel.ViewModelProviderFactory
+import dagger.android.support.DaggerFragment
+import kotlinx.android.synthetic.main.fragment_noun_home.*
+import javax.inject.Inject
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -22,15 +25,21 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [VerbHomeFragment.newInstance] factory method to
+ * Use the [NounHomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class VerbHomeFragment : Fragment(), VerbAdapter.OnVerbClickListener {
+class NounHomeFragment : DaggerFragment(), NounAdapter.OnNounClickListener {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var verbViewModel: VerbViewModel
+    private lateinit var nounViewModel: NounViewModel
+
+    @Inject
+    lateinit var providerFactory: ViewModelProviderFactory
+
+    @Inject
+    lateinit var nounAdapter: NounAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -45,38 +54,37 @@ class VerbHomeFragment : Fragment(), VerbAdapter.OnVerbClickListener {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_verb_home, container, false)
+        return inflater.inflate(R.layout.fragment_noun_home, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         initViewModels()
         initRecyclerView()
+
     }
 
     private fun initViewModels() {
-        verbViewModel = ViewModelProvider(this).get(VerbViewModel::class.java)
+        nounViewModel = ViewModelProvider(this, providerFactory).get(NounViewModel::class.java)
     }
 
     private fun initRecyclerView() {
-        val adapter = VerbAdapter(context!!, verbViewModel, viewLifecycleOwner, this)
-        recycler_verbhome.layoutManager = LinearLayoutManager(context!!)
 
-        recycler_verbhome.adapter = adapter
+        recycler_nounhome.layoutManager = LinearLayoutManager(context)
+        recycler_nounhome.adapter = nounAdapter
 
         var itemTouchHelper = ItemTouchHelper(
-            VerbSwipeToDeleteHelper(
+            NounSwipeToDeleteHelper(
                 this,
-                verbViewModel,
-                adapter, 0, ItemTouchHelper.RIGHT
+                nounViewModel,
+                nounAdapter, 0, ItemTouchHelper.RIGHT
             )
         )
 
-        itemTouchHelper.attachToRecyclerView(recycler_verbhome)
-        adapter.notifyDataSetChanged()
+        itemTouchHelper.attachToRecyclerView(recycler_nounhome)
+        nounAdapter.notifyDataSetChanged()
 
     }
-
 
     companion object {
         /**
@@ -85,12 +93,12 @@ class VerbHomeFragment : Fragment(), VerbAdapter.OnVerbClickListener {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment VerbListFragment.
+         * @return A new instance of fragment NounListFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            VerbHomeFragment().apply {
+            NounHomeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
@@ -98,7 +106,7 @@ class VerbHomeFragment : Fragment(), VerbAdapter.OnVerbClickListener {
             }
     }
 
-    override fun onVerbClick(id: Int) {
-
+    override fun onNounClick(id: Int) {
+        TODO("Not yet implemented")
     }
 }
