+++ app/src/main/java/com/teamzmron/selfstudyapp/Fragments/verb/VerbAddFragment.kt
@@ -1,4 +1,4 @@
-package com.teamzmron.selfstudyapp.Fragments
+package com.teamzmron.selfstudyapp.Fragments.verb
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -11,9 +11,11 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.teamzmron.selfstudyapp.Helper.Utils
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.Room.Entity.Adjective
-import com.teamzmron.selfstudyapp.ViewModel.AdjectiveViewModel
-import kotlinx.android.synthetic.main.fragment_adjective.*
+import com.teamzmron.selfstudyapp.Room.Entity.Verb
+import com.teamzmron.selfstudyapp.ViewModel.VerbViewModel
+import dagger.android.support.DaggerFragment
+import kotlinx.android.synthetic.main.fragment_verb.*
+
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -22,15 +24,15 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AdjectiveFragment.newInstance] factory method to
+ * Use the [VerbFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AdjectiveFragment : Fragment() {
+class VerbFragment : DaggerFragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var adjectiveViewModel: AdjectiveViewModel
+    private lateinit var verbViewModel: VerbViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -45,66 +47,67 @@ class AdjectiveFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_adjective, container, false)
+        return inflater.inflate(R.layout.fragment_verb, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
 
-        adjectiveViewModel = ViewModelProvider(this).get(AdjectiveViewModel::class.java)
+        verbViewModel = ViewModelProvider(this).get(VerbViewModel::class.java)
 
-        radio_i.isChecked = true
+        radioButton_verb_u.isChecked = true
 
-        button_adj_save.setOnClickListener {
+        button_verb_save.setOnClickListener {
             if (checkTextField()) {
-                saveInputFromLayout()
+                saveInputFromlayout()
             } else {
                 Toast.makeText(context!!, "All fields are required!", Toast.LENGTH_SHORT).show()
             }
+
         }
 
-        button_adj_clear.setOnClickListener {
+        button_verb_clear.setOnClickListener {
             clearTextFields()
         }
-
     }
 
-    private fun saveInputFromLayout() {
-        val adjType = if (getAdjType().contains("な")) "na" else "i"
-        val adjective = Adjective(
-            adjType = adjType,
-            adjNegative = editText_adj_negative.text.toString(),
-            adjPast = editText_adj_past.text.toString(),
-            adjPastNegative = editText_adj_pastnegative.text.toString(),
-            englishWord = editText_adj_englishWord.text.toString(),
-            japaneseWord = editText_adj_japaneseWord.text.toString(),
-            hiraganaForm = editText_adj_hirakata.text.toString(),
-            kanjiForm = editText_adj_kanji.text.toString(),
+    private fun saveInputFromlayout() {
+        val verbType = if (getVerbType().contains("")) "u" else "ru"
+        val verb = Verb(
+            verbType = verbType,
+            masu = editText_verb_masu.text.toString(),
+            masuPast = editText_verb_masupast.text.toString(),
+            masuNegative = editText_verb_masunegative.text.toString(),
+            masuPastNegative = editText_verb_masupastnegative.text.toString(),
+            englishWord = editText_verb_englishWord.text.toString(),
+            japaneseWord = editText_verb_japaneseWord.text.toString(),
+            hiraganaForm = editText_verb_hirakata.text.toString(),
+            kanjiForm = editText_verb_kanji.text.toString(),
             currentTimestamp = Utils.getTimeStamp()
         )
-
-        adjectiveViewModel.saveToDB(adjective)
-            .observe(viewLifecycleOwner, androidx.lifecycle.Observer {
-                if (it > 0) {
-                    clearTextFields()
-                    Toast.makeText(context!!, "Successfully added!", Toast.LENGTH_LONG).show()
-                }
-            })
+        verbViewModel.saveToDB(verb).observe(viewLifecycleOwner, androidx.lifecycle.Observer {
+            if (it > 0) {
+                clearTextFields()
+                Toast.makeText(context!!, "Successfully added!", Toast.LENGTH_LONG).show()
+            }
+        })
     }
 
-    private fun getAdjType(): String {
-        return radioGroup_adj.findViewById<RadioButton>(radioGroup_adj.checkedRadioButtonId).text.toString()
+    private fun getVerbType(): String {
+        return radioGroup_verb.findViewById<RadioButton>(radioGroup_verb.checkedRadioButtonId)
+            .text.toString()
     }
 
     private fun getAllTextFields(): ArrayList<EditText> {
         return arrayListOf(
-            editText_adj_negative,
-            editText_adj_past,
-            editText_adj_pastnegative,
-            editText_adj_englishWord,
-            editText_adj_japaneseWord,
-            editText_adj_hirakata,
-            editText_adj_kanji
+            editText_verb_masu,
+            editText_verb_masupast,
+            editText_verb_masunegative,
+            editText_verb_masupastnegative,
+            editText_verb_englishWord,
+            editText_verb_japaneseWord,
+            editText_verb_hirakata,
+            editText_verb_kanji
         )
     }
 
@@ -137,12 +140,12 @@ class AdjectiveFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment AdjectiveFragment.
+         * @return A new instance of fragment VerbFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AdjectiveFragment().apply {
+            VerbFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
