+++ app/src/main/java/com/teamzmron/selfstudyapp/Fragments/verb/VerbHomeFragment.kt
@@ -1,19 +1,22 @@
-package com.teamzmron.selfstudyapp.Fragments
+package com.teamzmron.selfstudyapp.Fragments.verb
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.teamzmron.selfstudyapp.Adapters.AdjectiveAdapter
-import com.teamzmron.selfstudyapp.Helper.AdjectiveSwipeToDeleteHelper
-
+import com.teamzmron.selfstudyapp.Adapters.VerbAdapter
+import com.teamzmron.selfstudyapp.Helper.VerbSwipeToDeleteHelper
 import com.teamzmron.selfstudyapp.R
-import com.teamzmron.selfstudyapp.ViewModel.AdjectiveViewModel
-import kotlinx.android.synthetic.main.fragment_adjective_home.*
+import com.teamzmron.selfstudyapp.ViewModel.VerbViewModel
+import com.teamzmron.selfstudyapp.ViewModel.ViewModelProviderFactory
+import dagger.android.support.DaggerFragment
+import kotlinx.android.synthetic.main.fragment_verb_home.*
+import javax.inject.Inject
+
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -22,16 +25,21 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AdjectiveHomeFragment.newInstance] factory method to
+ * Use the [VerbHomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AdjectiveHomeFragment : Fragment(), AdjectiveAdapter.OnAdjectiveClickListener {
+class VerbHomeFragment : DaggerFragment(), VerbAdapter.OnVerbClickListener {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
 
-    private lateinit var adjectiveViewModel: AdjectiveViewModel
+    private lateinit var verbViewModel: VerbViewModel
 
+    @Inject
+    lateinit var verbAdapter: VerbAdapter
+
+    @Inject
+    lateinit var providerFactory: ViewModelProviderFactory
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -46,7 +54,7 @@ class AdjectiveHomeFragment : Fragment(), AdjectiveAdapter.OnAdjectiveClickListe
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_adjective_home, container, false)
+        return inflater.inflate(R.layout.fragment_verb_home, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
@@ -56,25 +64,27 @@ class AdjectiveHomeFragment : Fragment(), AdjectiveAdapter.OnAdjectiveClickListe
     }
 
     private fun initViewModels() {
-        adjectiveViewModel = ViewModelProvider(this).get(AdjectiveViewModel::class.java)
+        verbViewModel = ViewModelProvider(this, providerFactory).get(VerbViewModel::class.java)
     }
 
     private fun initRecyclerView() {
-        val adapter = AdjectiveAdapter(context!!, adjectiveViewModel, viewLifecycleOwner, this)
-        recycler_adjhome.layoutManager = LinearLayoutManager(context!!)
-        recycler_adjhome.adapter = adapter
+        recycler_verbhome.layoutManager = LinearLayoutManager(context)
+
+        recycler_verbhome.adapter = verbAdapter
 
         var itemTouchHelper = ItemTouchHelper(
-            AdjectiveSwipeToDeleteHelper(
+            VerbSwipeToDeleteHelper(
                 this,
-                adjectiveViewModel,
-                adapter, 0, ItemTouchHelper.RIGHT
+                verbViewModel,
+                verbAdapter, 0, ItemTouchHelper.RIGHT
             )
         )
 
-        itemTouchHelper.attachToRecyclerView(recycler_adjhome)
-        adapter.notifyDataSetChanged()
+        itemTouchHelper.attachToRecyclerView(recycler_verbhome)
+        verbAdapter.notifyDataSetChanged()
+
     }
+
 
     companion object {
         /**
@@ -83,12 +93,12 @@ class AdjectiveHomeFragment : Fragment(), AdjectiveAdapter.OnAdjectiveClickListe
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment AdjectiveListFragment.
+         * @return A new instance of fragment VerbListFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AdjectiveHomeFragment().apply {
+            VerbHomeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
@@ -96,7 +106,7 @@ class AdjectiveHomeFragment : Fragment(), AdjectiveAdapter.OnAdjectiveClickListe
             }
     }
 
-    override fun onAdjectiveClick(id: Int) {
+    override fun onVerbClick(id: Int) {
 
     }
 }
