+++ app/src/main/java/com/teamzmron/selfstudyapp/ViewModel/NounViewModel.kt
@@ -3,38 +3,36 @@ package com.teamzmron.selfstudyapp.ViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import com.teamzmron.selfstudyapp.Repository.NounRepository
+import com.teamzmron.selfstudyapp.Room.Database.WordDatabase
 import com.teamzmron.selfstudyapp.Room.Entity.Noun
+import javax.inject.Inject
 
-class NounViewModel : ViewModel() {
-
-    private fun getNounRepoInstance() : NounRepository {
-        return NounRepository().getNounRepositoryInstance()
-    }
+class NounViewModel @Inject constructor(private val nounRepository: NounRepository) : ViewModel() {
 
     fun getWordsFromRepo(): LiveData<List<Noun>> {
-        return getNounRepoInstance().getNounFromDB()
+        return nounRepository.getNounFromDB()
     }
 
     fun getWordById(id: Int): LiveData<Noun> {
-        return getNounRepoInstance().getNounByIdFromDB(id)
+        return nounRepository.getNounByIdFromDB(id)
     }
 
     fun saveToDB(noun: Noun) : LiveData<Long> {
-        return getNounRepoInstance().saveNounRepo(noun)
+        return nounRepository.saveNounRepo(noun)
     }
 
     fun updateWord(noun: Noun) {
-        getNounRepoInstance().updateNounRepo(noun)
+        nounRepository.updateNounRepo(noun)
     }
 
 
     fun deleteWordById(noun: Noun) {
-        getNounRepoInstance().deleteNounRepo(noun)
+        nounRepository.deleteNounRepo(noun)
     }
 
 
     override fun onCleared() {
-        getNounRepoInstance().onClearDisposable()
+        nounRepository.onClearDisposable()
         super.onCleared()
     }
