--- app/src/main/java/com/teamzmron/selfstudyapp/di/ActivityBuildersModule.kt
@@ -1,17 +1,0 @@
-package com.teamzmron.selfstudyapp.di
-
-import com.teamzmron.selfstudyapp.HomeActivity
-import com.teamzmron.selfstudyapp.di.home.FragmentsModule
-import com.teamzmron.selfstudyapp.di.home.HomeModule
-import com.teamzmron.selfstudyapp.di.home.HomeViewModelsModule
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-
-@Module
-abstract class ActivityBuildersModule {
-
-    @ContributesAndroidInjector (
-        modules = [FragmentsModule::class, HomeViewModelsModule::class, HomeModule::class]
-    )
-    abstract fun contributeHomeActivity() : HomeActivity
-}
