+++ app/src/main/java/com/teamzmron/selfstudyapp/di/app/AppModule.kt
@@ -1,13 +1,15 @@
-package com.teamzmron.selfstudyapp.di
+package com.teamzmron.selfstudyapp.di.app
 
 import android.app.Application
 import com.teamzmron.selfstudyapp.Room.Database.WordDatabase
 import dagger.Module
 import dagger.Provides
+import javax.inject.Singleton
 
 @Module
 class AppModule {
 
+    @Singleton
     @Provides
     fun provideRoomDatabase(application: Application) : WordDatabase {
         return WordDatabase.getDatabasenIstance(application)
