--- app/src/main/java/com/teamzmron/selfstudyapp/di/home/HomeModule.kt
@@ -1,24 +1,0 @@
-package com.teamzmron.selfstudyapp.di.home
-
-import com.teamzmron.selfstudyapp.Adapters.AdjectiveAdapter
-import com.teamzmron.selfstudyapp.Adapters.NounAdapter
-import com.teamzmron.selfstudyapp.Adapters.VerbAdapter
-import dagger.Module
-import dagger.Provides
-
-@Module
-class HomeModule {
-
-    @HomeScope
-    @Provides
-    fun provideNounAdapter() : NounAdapter = NounAdapter()
-
-    @HomeScope
-    @Provides
-    fun provideAdjectiveAdapter() : AdjectiveAdapter = AdjectiveAdapter()
-
-    @HomeScope
-    @Provides
-    fun provideVerbAdapter() : VerbAdapter = VerbAdapter()
-
-}
