+++ app/src/main/java/com/teamzmron/selfstudyapp/di/main/MainModule.kt
@@ -1,0 +1,42 @@
+package com.teamzmron.selfstudyapp.di.main
+
+import com.teamzmron.selfstudyapp.Adapters.AdjectiveAdapter
+import com.teamzmron.selfstudyapp.Adapters.NounAdapter
+import com.teamzmron.selfstudyapp.Adapters.VerbAdapter
+import com.teamzmron.selfstudyapp.Repository.AdjectiveRepository
+import com.teamzmron.selfstudyapp.Repository.NounRepository
+import com.teamzmron.selfstudyapp.Repository.VerbRepository
+import com.teamzmron.selfstudyapp.Room.Database.WordDatabase
+import dagger.Module
+import dagger.Provides
+
+
+@Module
+class MainModule {
+
+    @MainScope
+    @Provides
+    fun provideNounAdapter() : NounAdapter = NounAdapter()
+
+    @MainScope
+    @Provides
+    fun provideAdjectiveAdapter() : AdjectiveAdapter = AdjectiveAdapter()
+
+    @MainScope
+    @Provides
+    fun provideVerbAdapter() : VerbAdapter = VerbAdapter()
+
+    @MainScope
+    @Provides
+    fun verbRepository(wordDatabase: WordDatabase) : VerbRepository = VerbRepository(wordDatabase)
+
+    @MainScope
+    @Provides
+    fun nounRepository(wordDatabase: WordDatabase) : NounRepository = NounRepository(wordDatabase)
+
+    @MainScope
+    @Provides
+    fun adjRepository(wordDatabase: WordDatabase) : AdjectiveRepository = AdjectiveRepository(wordDatabase)
+
+
+}
