+++ app/src/main/java/com/teamzmron/selfstudyapp/di/main/MainScope.kt
@@ -1,4 +1,4 @@
-package com.teamzmron.selfstudyapp.di.home
+package com.teamzmron.selfstudyapp.di.main
 
 import java.lang.annotation.Documented
 import javax.inject.Scope
@@ -6,5 +6,5 @@ import javax.inject.Scope
 @Scope
 @Documented
 @Retention(AnnotationRetention.RUNTIME)
-annotation class HomeScope {
+annotation class MainScope {
 }
