+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    android:id="@+id/drawer_layout">
+    android:id="@+id/drawer">
 
 
     <fragment
@@ -27,7 +27,7 @@
             android:layout_height="wrap_content"
             style="@style/Widget.AppCompat.ProgressBar"
             android:layout_centerInParent="true"
-            android:id="@+id/progress_bar"
+            android:id="@+id/progressBar"
             android:visibility="gone"
             android:elevation="20dp"/>
 
@@ -35,7 +35,7 @@
 
 
     <com.google.android.material.navigation.NavigationView
-        android:id="@+id/nav_view"
+        android:id="@+id/navView"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
         android:layout_gravity="start"
