+++ app/src/main/res/navigation/main.xml
@@ -6,24 +6,24 @@
     app:startDestination="@+id/homeFragment">
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.teamzmron.selfstudyapp.Fragments.NounHomeFragment"
+        android:name="com.teamzmron.selfstudyapp.Fragments.Home"
         android:label="@string/nounHomeList"
-        tools:layout="@layout/fragment_noun_home" />
+        tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/nounList"
-        android:name="com.teamzmron.selfstudyapp.Fragments.NounHomeFragment"
+        android:name="com.teamzmron.selfstudyapp.Fragments.noun.NounHomeFragment"
         android:label="@string/nounHomeList"
         tools:layout="@layout/fragment_noun_home" />
 
     <fragment
         android:id="@+id/verbList"
-        android:name="com.teamzmron.selfstudyapp.Fragments.VerbHomeFragment"
+        android:name="com.teamzmron.selfstudyapp.Fragments.verb.VerbHomeFragment"
         android:label="@string/verbHomeList"
         tools:layout="@layout/fragment_verb_home" />
 
     <fragment
         android:id="@+id/adjList"
-        android:name="com.teamzmron.selfstudyapp.Fragments.AdjectiveHomeFragment"
+        android:name="com.teamzmron.selfstudyapp.Fragments.Adjective.AdjectiveHomeFragment"
         android:label="@string/adjHomeList"
         tools:layout="@layout/fragment_adjective_home" />
