+++ app/src/main/res/values/strings.xml
@@ -8,5 +8,6 @@
     <string name="noun">Noun</string>
     <string name="verb">Verb</string>
     <string name="adjective">Adjective</string>
+    <string name="home">Home</string>
 
 </resources>
